fix: Squashed migrations
