--- alembic/versions/6e3d14bbe099_country.py
@@ -1,42 +1,0 @@
-"""country
-
-Revision ID: 6e3d14bbe099
-Revises: 622211feaeb0
-Create Date: 2023-04-12 12:39:01.805183
-
-"""
-import sqlalchemy as sa
-from alembic import op
-
-__all__ = ["downgrade", "upgrade"]
-
-
-# revision identifiers, used by Alembic.
-revision = "6e3d14bbe099"
-down_revision = "622211feaeb0"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "country",
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("population", sa.Integer(), nullable=False),
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column("created", sa.DateTime(), nullable=False),
-        sa.Column("updated", sa.DateTime(), nullable=False),
-        sa.PrimaryKeyConstraint("id", name=op.f("pk_country")),
-    )
-    op.add_column("author", sa.Column("country_id", sa.Uuid(), nullable=True))
-    op.create_foreign_key(op.f("fk_author_country_id_country"), "author", "country", ["country_id"], ["id"])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f("fk_author_country_id_country"), "author", type_="foreignkey")
-    op.drop_column("author", "country_id")
-    op.drop_table("country")
-    # ### end Alembic commands ###
