+++ alembic/versions/c41960c217ac_squash.py
@@ -1,0 +1,53 @@
+"""squash
+
+Revision ID: c41960c217ac
+Revises:
+Create Date: 2023-07-06 03:46:54.476229
+
+"""
+import litestar
+import sqlalchemy as sa
+from alembic import op
+
+__all__ = ["downgrade", "upgrade"]
+
+
+# revision identifiers, used by Alembic.
+revision = "c41960c217ac"
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "country",
+        sa.Column("name", sa.String(), nullable=False),
+        sa.Column("population", sa.Integer(), nullable=False),
+        sa.Column("id", litestar.contrib.sqlalchemy.types.GUID(length=16), nullable=False),
+        sa.Column("_sentinel", sa.Integer(), nullable=True),
+        sa.Column("created_at", litestar.contrib.sqlalchemy.types.DateTimeUTC(timezone=True), nullable=False),
+        sa.Column("updated_at", litestar.contrib.sqlalchemy.types.DateTimeUTC(timezone=True), nullable=False),
+        sa.PrimaryKeyConstraint("id", name=op.f("pk_country")),
+    )
+    op.create_table(
+        "author",
+        sa.Column("name", sa.String(), nullable=False),
+        sa.Column("dob", sa.Date(), nullable=False),
+        sa.Column("country_id", litestar.contrib.sqlalchemy.types.GUID(length=16), nullable=True),
+        sa.Column("id", litestar.contrib.sqlalchemy.types.GUID(length=16), nullable=False),
+        sa.Column("_sentinel", sa.Integer(), nullable=True),
+        sa.Column("created_at", litestar.contrib.sqlalchemy.types.DateTimeUTC(timezone=True), nullable=False),
+        sa.Column("updated_at", litestar.contrib.sqlalchemy.types.DateTimeUTC(timezone=True), nullable=False),
+        sa.ForeignKeyConstraint(["country_id"], ["country.id"], name=op.f("fk_author_country_id_country")),
+        sa.PrimaryKeyConstraint("id", name=op.f("pk_author")),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("author")
+    op.drop_table("country")
+    # ### end Alembic commands ###
